+++ crates/ludtwig-parser/src/lexer.rs
@@ -304,6 +304,11 @@ mod tests {
         add("€", T![unknown]);
         add("trans", T!["trans"]);
         add("endtrans", T!["endtrans"]);
+        add("sw_embed", T!["sw_embed"]);
+        add("sw_end_embed", T!["sw_end_embed"]);
+        add("sw_use", T!["sw_use"]);
+        add("sw_import", T!["sw_import"]);
+        add("sw_from", T!["sw_from"]);
 
         // lex and compare
         let results = lex(&source);
@@ -1082,5 +1087,30 @@ mod tests {
     #[test]
     fn lex_ludtwig_ignore() {
         check_token("ludtwig-ignore", T!["ludtwig-ignore"]);
+    }
+
+    #[test]
+    fn lex_sw_embed() {
+        check_token("sw_embed", T!["sw_embed"]);
+    }
+
+    #[test]
+    fn lex_sw_end_embed() {
+        check_token("sw_end_embed", T!["sw_end_embed"]);
+    }
+
+    #[test]
+    fn lex_sw_use() {
+        check_token("sw_use", T!["sw_use"]);
+    }
+
+    #[test]
+    fn lex_sw_import() {
+        check_token("sw_import", T!["sw_import"]);
+    }
+
+    #[test]
+    fn lex_sw_from() {
+        check_token("sw_from", T!["sw_from"]);
     }
 }